Weather done

+++ app/src/main/java/com/renegade/weas/network/response/weatherresponse/Wind.kt
@@ -1,6 +1,6 @@
 package com.renegade.weas.network.response.weatherresponse
 
 data class Wind(
-    val deg: Int,
+    val deg: Double,
     val speed: Double
 )
